refactor: extract JointImpWrapper and IkChainHelper from JointHierarchyManager
Extract two package-private classes to reduce JointHierarchyManager from 650 to 346 lines:

- JointImpWrapper (244 lines): inner class promoted to top-level with
  constructor-injected JointedModelResourceBinder<?> replacing outer-class
  reference for isSims() access.

- IkChainHelper (193 lines): AddOp enum + 3 IK chain methods extracted as
  static utilities parameterized with Function<JointId, JointImp> for joint
  lookup, decoupling from the hierarchy manager's internal map.

JointHierarchyManager retains hierarchy construction, joint lookup, tree walk,
caches, JointData, and StraightenTreeWalkObserver. Delegates IK computation
to IkChainHelper via method reference (this::getJointImplementation).

No public API changes. All 27 new + 46 existing decomposition tests pass.

Closes #641
